refactor(copilot-poller): extract createCopilotStateDb to shared test-helpers; fix docstring source reference
- Extract duplicate createStateDb from copilot-poller.spec.ts and
  copilot-poller-integration.spec.ts into packages/daemon/src/github/
  test-helpers.ts as createCopilotStateDb (addresses review thread #1793)
- Update file header docstring: 'documented from API observation' →
  'documented from manual API observation, captured in PR #1793 body'
  (addresses review thread #1793)
